Separated out Property Gateway and Property Gateway Children tests into different files.
